tutorial/pyproject.toml
[project]
name = "poc-ai-agent-learning"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "duckduckgo-search>=8.1.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-openai>=0.3.31",
    "langgraph>=0.6.6",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
]
